Adapt application-password validator to wordpress-rs UInt status codes
The wordpress-rs bump that landed on trunk (fc05d3bb3d9) changes
WpError.statusCode / UnknownError.statusCode / InvalidHttpStatusCode.statusCode
from `UShort` to `UInt`. Update the validator's `isAuthStatusCode`
parameter and the test fixtures accordingly. No behavior change.

+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/cards/applicationpassword/ApplicationPasswordValidator.kt
@@ -86,7 +86,7 @@ class ApplicationPasswordValidator @Inject constructor(
             code is WpErrorCode.ApplicationPasswordNotFound ||
             code is WpErrorCode.NoAuthenticatedAppPassword
 
-    private fun isAuthStatusCode(statusCode: UShort): Boolean =
+    private fun isAuthStatusCode(statusCode: UInt): Boolean =
         statusCode.toInt() == HTTP_UNAUTHORIZED || statusCode.toInt() == HTTP_FORBIDDEN
 
     private fun isAuthErrorReason(reason: RequestExecutionErrorReason): Boolean =
